+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   * Bugfix: Ensure that failure to dial a call kills the actor
   * Bugfix: Adhearsion call actors should do recursion detection on #inspect
   * Bugfix: Fix controller block context on JRuby
+  * Bugfix: Require only Bundler groups appropriate to application environment
 
 # [2.4.0](https://github.com/adhearsion/adhearsion/compare/v2.3.5...v2.4.0) - [2013-08-29](https://rubygems.org/gems/adhearsion/versions/2.4.0)
   * Deprecation: Ruby 1.9.2 support is deprecated and will be dropped in a future version of Adhearsion